fix: 重载资源导致 audio soruce失效

+++ core/src/main/java/top/tobyprime/mcedia/core/MediaPlayer.java
@@ -131,7 +131,7 @@ public class MediaPlayer {
             media.pause();
         }
     }
-    public synchronized void seek(int ms){
+    public synchronized void seek(long ms){
         if (media != null) {
             media.seek(ms);
         }
